feat : 적 대미지 다르게 들어가도록 scriptableObjects

+++ Assets/Scripts/interface/IDamageable.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void Damage(float damage);
+    public void Damage(int damage);
     
 }
